+++ src/Main.hs
@@ -1,22 +1,33 @@
 module Main where
 
+import System.Posix.Types (FileOffset)
 import System.Posix.Files (getFileStatus, fileSize)
 import System.Directory (doesDirectoryExist, listDirectory)
 import Control.Monad (forM)
 import System.FilePath.Posix ((</>))
 import System.Environment (getArgs)
+import Data.List (sortBy)
+import Data.Function (on)
 
+-- pure
+mySort :: Ord b => [(a, b)] -> [(a, b)]
+mySort = sortBy (flip compare `on` snd)
+
+-- impure
+getFileSize :: FilePath -> IO FileOffset
 getFileSize filepath = do
   status <- getFileStatus filepath
   return (fileSize status)
 
+completePath :: FilePath -> IO [FilePath]
 completePath mainDir = do
   simpleFileNames <- listDirectory mainDir
   allPaths <- forM simpleFileNames $ \simpleName -> do
     let path = mainDir </> simpleName
     return [path]
   return (concat allPaths)
-  
+
+getSize :: FilePath -> IO [FileOffset]
 getSize filepath = do
   isDir <- doesDirectoryExist filepath
   filesize0 <- getFileSize filepath
@@ -29,17 +40,15 @@ getSize filepath = do
       if isDirInside
         then getSize filename
         else return [filesize1]
-        -- with filename
-        --else return [(filename, filesize1)]
     return (concat allLocalPaths)
     else return [filesize0]
-    -- with filename
-    -- else return [(filepath, filesize0)]
 
+getSizeSum :: FilePath -> IO FileOffset
 getSizeSum filepath = do
   filesizes <- getSize filepath
   return (sum filesizes)
 
+listGreater :: FilePath -> IO [(FilePath, FileOffset)]
 listGreater filepath = do
   isDir <- doesDirectoryExist filepath
   filesize0 <- getFileSize filepath
@@ -49,9 +58,10 @@ listGreater filepath = do
     everyFile <- forM localFiles $ \filename -> do
       sumfile <- getSizeSum filename
       return (filename, sumfile)
-    return everyFile
+    return (mySort everyFile)
     else return [(filepath, filesize0)]
 
+main :: IO ()
 main = do
   (filenam:_) <- getArgs
   listgreater <- listGreater filenam